Remove flag to refresh wikidata, and pick a compromise expiration
Setting expire_after to 0 didn't work as I expected. For now, I just
picked a compromise between caching the responses for longer and not
waiting too long for changes to propagate. But if
https://github.com/requests-cache/requests-cache/issues/899 is
implemented, it would be nice to go back to having a longer expiration
with a flag to revalidate.