Don't ignore children when the parent-child border isn't crossed
If the parent is being processed, this shouldn't change anything because
the child wouldn't be processed either way, whether because it's
integral or because the border isn't crossed.

If the child is being processed, the old behavior was that the parent
wasn't processed because of the border, but the child was still ignored
for being integral. With this change, the child is no longer ignored in
that case.